Durable broadcast log + multi-host fan-out and stream failover.
Broadcast: new mochi.broadcast.send and mochi.broadcast.replay APIs, per-app _log/_acknowledged/_sequence/_received tables (renamed from _broadcast_*), gap detection wrapper in events.go that NACKs out-of-order events and fires async resync, system handlers for broadcast/resync and broadcast/acknowledge. Inline cleanup: 7-day age cap on send, ack-driven trim on broadcast/acknowledge. No scheduled tasks.

Multi-host fan-out: messages.send fans out to every active directory location for the target entity (was: single peer). queue rows expand on demand when an entity gets directory entries after enqueue.

Stream/RPC failover: entity_peers_failover returns active peers (seen within 2h) oldest-first then stale peers most-recent first; stream() and remote_connect() try each until one connects so a single-host outage doesn't break delivery.

Tests: broadcast_log_test.go covers log append, per-peer sequences, age trim, ack trim, replay, resync throttle. entities_failover_test.go covers active/stale tiering, aged-out exclusion, every-peer-included guarantee. messages_fanout_test.go covers fan-out, unknown-entity retry, send_peer pinned-target, queue expansion on late directory resolution.

+++ server/replication_test.go
@@ -1083,7 +1083,7 @@ func TestBroadcastGapDetection(t *testing.T) {
 
 	// Sequence 5 arrives — gap of {3, 4} detected (app would request
 	// replay; we just check the math).
-	incoming := 5
+	incoming := int64(5)
 	gap := incoming > last+1
 	if !gap {
 		t.Errorf("gap should be detected when incoming > last+1")
